use helper class; add copyright headers
